examples/python/clients/solana/pyproject.toml
[project]
name = "x402-solana-client-example"
version = "0.1.0"
description = "Example x402 client for Solana payments"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "x402>=0.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

